+++ TablePro/Views/Main/MainContentView.swift
@@ -329,7 +329,6 @@ struct MainContentView: View {
                 setupCommandActions()
                 updateToolbarPendingState()
                 updateInspectorContext()
-                coordinator.aiViewModel = rightPanelState.aiViewModel
                 coordinator.rightPanelState = rightPanelState
 
                 Self.lifecycleLogger.info(